+++ lib/screen/home/home_screen.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:ohayo_post_app/counter_notifier.dart';
-import 'package:ohayo_post_app/firebase_notifier.dart';
-import 'package:ohayo_post_app/login_screen.dart';
-import 'package:ohayo_post_app/registration_screen.dart';
+import 'package:ohayo_post_app/notifier/counter_notifier.dart';
+import 'package:ohayo_post_app/notifier/firebase_notifier.dart';
+import 'package:ohayo_post_app/screen/login/login_screen.dart';
+import 'package:ohayo_post_app/screen/registration/registration_screen.dart';
 import 'package:provider/provider.dart';
 
 class HomeScreen extends StatelessWidget {
